now there are two implementations of PlumeNavigationTask. I merged them and used the one that Alfred has updated (ie with the zoom-in camera)
